refactor: move save to localStorage middleware to store
