+++ src/ducks/store/configureStore.prod.js
@@ -2,7 +2,7 @@ import { applyMiddleware, createStore, compose } from "redux";
 import createSagaMiddleware from "redux-saga";
 import rootSaga from "../rootSaga";
 import rootReducer from "../rootReducer";
-import saveToStorage from "../../utils/saveStateToLocalStorage";
+import saveToStorage from "./utils";
 
 const sagaMiddleware = createSagaMiddleware();
 const middlewares = applyMiddleware(sagaMiddleware);
